requirements.txt
sounddevice==0.4.7
PySide6==6.5.2
PyYAML==6.0.1
numpy==1.26.4
faster-whisper==1.0.2
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cublas-cu12==12.2.5.6
nvidia-cudnn-cu12==8.9.7.29